Fixed things
I decided to change some stuff around. I got rid of keeping arrays in txt files because that wasn't a good idea, at least at my current skill level. I also got it so that I can add a random bit to the end of one of the prompts, which'll be important for the future.