Smaller one
I added a temp placeholder script for the value evaluation of prompts. I made it so that it keeps everything in the array, and sends that off to the new script which is basically just the save conversation script from the end at this point. Also decided to remove thing.py.