Finally.
I haven't tested it yet and I know things are probably broken or messed up or unfinished. I went through everything and when I fix whatever I broke it'll be better. Didn't touch the plain text stuff or the saves stuff, removed main.py because this projects not ready for a UI yet, yeah.