ReadMe Fixed!!!
This is a big one. I improved the readme heaps and bounds in my eyes, and several other changes. This took me all day, I'm happy with the results. Sorry it's so long. I just pasted it into grammarly and I have no idea how I let myself write 5, now closer to 4 pages of text for the readme!