Planning and whatnot
Moved a lot of planning + project details to their own .md file + edited it a bit, wrote in some notes in a few files, and made an empty py file with what I'm gonna eventually put in it. Yeah. Got a bit lost for a while on where I wanted to take this project, but I think I have more of an idea now and this is an update to add those ideas into the project, alongside other changes as always.