Start to update settings
Should work. It's in it's own branch because I thought I wouldn't have a published version by the time I was done working on this thing. It's also not really done and probably a good idea for me to start utilizing branches W official patches and updates posted on main for organization and actually working code stuff.